Phase 3.2: Split RemoteArtifactBoardController into sender + receiver
Extract RemoteBoardSyncSender (~310 LOC) and RemoteBoardSyncReceiver (~340 LOC)
from RemoteArtifactBoardController (1,209 → ~290 LOC).

- Sender: outbound SignalR pushes, debounce timers, size listeners, snapshots
- Receiver: inbound SignalR callbacks, artifact creation/update helpers
- Controller: thin orchestrator delegating to sender + receiver

All Flutter tests pass. Zero analyze errors.
